autocxx_parser
Module file_locations

Enums

FileLocationStrategy

The strategy used to generate, and to find generated, files. As standard, these are based off the OUT_DIR set by Cargo, but our code can’t assume it can read OUT_DIR as it may be running in the rust-analyzer proc macro server where it’s not available. We need to give provision for custom locations and we need our code generator build script to be able to pass locations through to the proc macro by env vars.